feat: making models to run in double precision, adding extra input args

+++ experiments/models/pretrained_mace.py
@@ -274,7 +274,7 @@ def return_mace_calc():
     model.encoder.encoder.load_state_dict(
         torch.load(ckpt, map_location=torch.device("cpu")).state_dict(), strict=True
     )
-
+    model = model.to(torch.double)
     transforms = [
         PeriodicPropertiesTransform(cutoff_radius=5.0, adaptive_cutoff=True),
         PointCloudToGraphTransform(
